renamed index to DotUptime (it was no index, but a composite DotUptime Analyzer
